Add Lexer stub, Regex code
Add stub code that will represent the Lexer API that is exposed to
external clients. The Regex set of classes represents an object
model for regular expressions.

The NFAGen class was moved from Regex/ to Gen/ to clearly mark what
is pure Regex representation code and actual NFA generation and
manipulation work.