Fix accept DFA reset bug
If peek() on the next character results in a potentially longer match,
then all currently accepting DFAs are reset.

However, if one of the currently accepting DFAs can successfully peek()
on the next character, then it *should not* be reset. Only reset the
accept DFAs that would reject the next character.

This fix allows for the star (*) regular expression operator to match
as large a string as possible, which is the expected (greedy) behavior.